+++ packages/react-core/src/components/Modal/ModalBoxCloseButton.tsx
@@ -1,4 +1,6 @@
 import * as React from 'react';
+import { css } from '@patternfly/react-styles';
+import styles from '@patternfly/react-styles/css/components/ModalBox/modal-box';
 import { Button } from '../Button';
 import TimesIcon from '@patternfly/react-icons/dist/esm/icons/times-icon';
 import { OUIAProps } from '../../helpers';
@@ -13,20 +15,21 @@ export interface ModalBoxCloseButtonProps extends OUIAProps {
 }
 
 export const ModalBoxCloseButton: React.FunctionComponent<ModalBoxCloseButtonProps> = ({
-  className = '',
+  className,
   onClose = () => undefined as any,
   ouiaId,
   ...props
 }: ModalBoxCloseButtonProps) => (
-  <Button
-    className={className}
-    variant="plain"
-    onClick={onClose}
-    aria-label="Close"
-    {...(ouiaId && { ouiaId: `${ouiaId}-${ModalBoxCloseButton.displayName}` })}
-    {...props}
-  >
-    <TimesIcon />
-  </Button>
+  <div className={css(styles.modalBoxClose, className)}>
+    <Button
+      variant="plain"
+      onClick={onClose}
+      aria-label="Close"
+      {...(ouiaId && { ouiaId: `${ouiaId}-${ModalBoxCloseButton.displayName}` })}
+      {...props}
+    >
+      <TimesIcon />
+    </Button>
+  </div>
 );
 ModalBoxCloseButton.displayName = 'ModalBoxCloseButton';