+++ packages/react-core/src/components/Toolbar/examples/ToolbarStacked.tsx
@@ -2,19 +2,17 @@ import React from 'react';
 import {
   Button,
   ButtonVariant,
-  KebabToggle,
   Select,
   SelectOption,
   SelectOptionObject,
   SelectVariant,
   Pagination,
   Dropdown,
-  DropdownSeparator,
   DropdownToggle,
   DropdownToggleCheckbox,
   DropdownItem,
-  DropdownPosition,
   Divider,
+  MenuToggle,
   OverflowMenu,
   OverflowMenuContent,
   OverflowMenuControl,
@@ -25,7 +23,9 @@ import {
   ToolbarToggleGroup,
   ToolbarItem
 } from '@patternfly/react-core';
+import { Dropdown as DropdownNext, DropdownItem as DropdownItemNext, DropdownList } from '@patternfly/react-core/next';
 import FilterIcon from '@patternfly/react-icons/dist/esm/icons/filter-icon';
+import EllipsisVIcon from '@patternfly/react-icons/dist/esm/icons/ellipsis-v-icon';
 
 export const ToolbarStacked: React.FunctionComponent = () => {
   // toggle group - three option menus with labels, two icon buttons, Kebab menu - right aligned
@@ -52,8 +52,8 @@ export const ToolbarStacked: React.FunctionComponent = () => {
   const [page, setPage] = React.useState(1);
   const [perPage, setPerPage] = React.useState(20);
 
-  const onKebabToggle = (_event: any, isOpen: boolean) => {
-    setKebabIsOpen(isOpen);
+  const onKebabToggle = () => {
+    setKebabIsOpen(!kebabIsOpen);
   };
 
   const onResourceToggle = (_event: any, isExpanded: boolean) => {
@@ -65,7 +65,7 @@ export const ToolbarStacked: React.FunctionComponent = () => {
     setResourceIsExpanded(false);
   };
 
-  const onResourceSelectDropdown = (event: React.SyntheticEvent<HTMLDivElement, Event> | undefined) => {
+  const onResourceSelectDropdown = (event: React.MouseEvent<Element, MouseEvent>| undefined) => {
     setResourceSelected(event?.target);
     setResourceIsExpanded(false);
   };
@@ -101,21 +101,21 @@ export const ToolbarStacked: React.FunctionComponent = () => {
   };
 
   const dropdownItems = [
-    <DropdownItem key="link">Link</DropdownItem>,
-    <DropdownItem key="action" component="button">
+    <DropdownItemNext key="link">Link</DropdownItemNext>,
+    <DropdownItemNext key="action" component="button">
       Action
-    </DropdownItem>,
-    <DropdownItem key="disabled link" isDisabled>
+    </DropdownItemNext>,
+    <DropdownItemNext key="disabled link" isDisabled>
       Disabled Link
-    </DropdownItem>,
-    <DropdownItem key="disabled action" isDisabled component="button">
+    </DropdownItemNext>,
+    <DropdownItemNext key="disabled action" isDisabled component="button">
       Disabled Action
-    </DropdownItem>,
-    <DropdownSeparator key="separator" />,
-    <DropdownItem key="separated link">Separated Link</DropdownItem>,
-    <DropdownItem key="separated action" component="button">
+    </DropdownItemNext>,
+    <Divider key="separator" />,
+    <DropdownItemNext key="separated link">Separated Link</DropdownItemNext>,
+    <DropdownItemNext key="separated action" component="button">
       Separated Action
-    </DropdownItem>
+    </DropdownItemNext>
   ];
 
   const splitButtonDropdownItems = [
@@ -192,14 +192,23 @@ export const ToolbarStacked: React.FunctionComponent = () => {
                 </OverflowMenuGroup>
               </OverflowMenuContent>
               <OverflowMenuControl hasAdditionalOptions>
-                <Dropdown
-                  onSelect={onResourceSelectDropdown}
-                  toggle={<KebabToggle onToggle={onKebabToggle} />}
-                  isOpen={kebabIsOpen}
-                  isPlain
-                  dropdownItems={dropdownItems}
-                  position={DropdownPosition.right}
-                />
+              <DropdownNext
+                   onSelect={onResourceSelectDropdown}
+                   toggle={toggleRef => (
+                     <MenuToggle
+                       ref={toggleRef}
+                       aria-label="Kebab overflow menu"
+                       variant="plain"
+                       onClick={onKebabToggle}
+                       isExpanded={kebabIsOpen}
+                     >
+                       <EllipsisVIcon />
+                     </MenuToggle>
+                   )}
+                   isOpen={kebabIsOpen}
+                 >
+                   <DropdownList>{dropdownItems}</DropdownList>
+                 </DropdownNext>
               </OverflowMenuControl>
             </OverflowMenu>
           </ToolbarItem>